Catch disabled packages in ApplicationsProvider.
ApplicationsProvider overrides PackageMonitor.onSomePackagesChanged
to watch for package changes. However, this isn't called when an
application is disabled

When GooglePartnerServices disables the "Books" app because the phone is
outside the US, ApplicationsProvider doesn't see this, so QSB still
shows the Books app. Override onPackageModified as well, and post an
update for that package, so we'll remove the app from the DB.

Bug: 5029707
Change-Id: I48f5a608f174b56544f52e379c80db309d2dff47